test: add tests for recurring transactions

+++ app/Livewire/RecurringTransactions/Index.php
@@ -4,13 +4,17 @@ declare(strict_types=1);
 
 namespace App\Livewire\RecurringTransactions;
 
+use App\Concerns\WithBreadcrumbs;
+use App\Data\BreadcrumbItemData;
 use App\Models\RecurringTransaction;
+use Flux\Flux;
 use Illuminate\Contracts\View\View;
 use Livewire\Component;
 use Livewire\WithPagination;
 
 final class Index extends Component
 {
+    use WithBreadcrumbs;
     use WithPagination;
 
     public function toggleStatus(RecurringTransaction $transaction): void
@@ -22,11 +26,16 @@ final class Index extends Component
 
     public function delete(RecurringTransaction $transaction): void
     {
+        $this->authorize('delete', $transaction);
+        Flux::toast(__('recurring_transactions.deleted', ['name' => $transaction->title]));
+        Flux::modal('confirm-deletion-'.$transaction->id)->close();
         $transaction->delete();
     }
 
     public function render(): View
     {
+        $this->withBreadcrumbs(new BreadcrumbItemData(__('recurring_transactions.index')));
+
         return view('livewire.recurring-transactions.index', [
             'transactions' => RecurringTransaction::with(['wallet', 'category'])
                 ->latest()
